backend/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
python-multipart>=0.0.9,<1.0
pydantic-settings>=2.3,<3.0
pillow>=10.0,<12.0
ultralytics>=8.2,<9.0
inference-sdk>=0.31,<1.0
opencv-python>=4.9,<5.0
numpy>=1.26,<3.0
